// src/main/webapp/app/shared/model/data-pipeline/pipeline.model.ts
/*
 * This Source Code Form is subject to the terms of the Mozilla Public License, v.
 * 2.0 with a Healthcare Disclaimer.
 * A copy of the Mozilla Public License, v. 2.0 with the Healthcare Disclaimer can
 * be found under the top level directory, named LICENSE.
 * If a copy of the MPL was not distributed with this file, You can obtain one at
 * http://mozilla.org/MPL/2.0/.
 * If a copy of the Healthcare Disclaimer was not distributed with this file, You
 * can obtain one at the project website https://github.com/igia.
 *
 * Copyright (C) 2018-2019 Persistent Systems, Inc.
 */
import { DataPipelineState } from './data-pipeline-state.model';
import { Endpoint } from '../endpoint/endpoint.model';
import { Audit } from '../audit/audit.model';
import { AuditMessageType } from './audit-message-type.model';

export interface Pipeline extends Audit {
    id?: number;
    name?: string;
    description?: string;
    deploy?: boolean;
    version?: string;
    state?: DataPipelineState;
    reason?: string;
    workerService?: string;
    auditMessages?: AuditMessageType[];
    source?: Endpoint;
    destinations?: Endpoint[];
}